Poprawka zadania 2

+++ lab03/lab03_02.c
@@ -15,8 +15,8 @@ int main()
     int iloczyn = a * b;
     printf("Iloczyn: %d\n", iloczyn);
 
-    int iloraz = a / b;
-    printf("Iloraz: %d\n", iloraz);
+    float iloraz = (float)a / b;
+    printf("Iloraz: %f\n", iloraz);
 
     int modulo = a % b;
     printf("Reszta z dzielenia: %d\n", modulo);
